chore: add end-to-end tracing to OCR pipeline
Instrument every stage boundary with element/text counts at debug level
to enable isolating content loss through the pipeline:
- hOCR parser: input bytes → element count + text chars
- ocr_doc_to_paragraphs: input elements → output paragraphs
- apply_layout_overrides: classification breakdown (headings, lists, etc.)
- assemble_internal_document: output element count
- OCR extraction: per-page paragraph count + raw content length
- PaddleOCR backend: text blocks → OCR elements → InternalDoc elements

+++ tests/test_apps/go/extraction_test.go
@@ -1784,7 +1784,7 @@ func TestMetadataPptxMetadata(t *testing.T) {
 	pptx, ok := metadata.PptxMetadata()
 	assert.True(t, ok)
 	assert.NotNil(t, pptx)
-	assert.Equal(t, uint64(3), pptx.SlideCount)
+	assert.Equal(t, 3, pptx.SlideCount)
 	assert.Equal(t, 3, len(pptx.SlideNames))
 }
